+++ src/types/InteractionType.ts
