+++ src/types/Interactions.ts
